Cleanup and remove old config logic
