--- sdk/core/src/main/java/com/klaviyo/coresdk/utils/ConfigKeys.kt
@@ -1,6 +1,0 @@
-package com.klaviyo.coresdk.utils
-
-internal object ConfigKeys {
-    const val UUID_KEY = "uuid"
-    const val PUSH_TOKEN_KEY = "device_token"
-}
